+++ api/utils/validator.js
@@ -14,5 +14,22 @@ module.exports = {
 
     validateCoordination: function(lat, lng) {
         return (lat >= -90 && lat <= 90) && (lng >= -180 && lng <= 180);
+    },
+
+    validateOpenHour: function(event, restaurant) {
+        var isValid = false;
+        var startTime = new Date(event.startTime);
+        var day = startTime.getDay();
+        var time = "" + startTime.getHours() + startTime.getMinutes();
+
+        restaurant.hours.forEach(function(element) {
+            element.open.forEach(function(entry){
+                if(day == entry.day){
+                    isValid = (time > entry.start && time < entry.end);
+                }
+            })
+        });
+
+        return isValid;
     }
 };
